# api/sarcoma.openapi.yaml
openapi: 3.0.3
info:
  title: Sarcom FastTrack API
  version: 1.0.0
  description: Go WebAPI for Sarcom FastTrack patient handoff reports.
  contact:
    email: support@sarcoma-fasttrack.local
  license:
    name: CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/
servers:
  - description: Local API
    url: http://localhost:8000
  - description: WAC Gateway API route
    url: /sarcoma-fasttrack-api
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/users:
    get:
      tags: [users]
      summary: List users.
      operationId: listUsers
      responses:
        "200":
          description: Visible users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRead"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [users]
      summary: Create a user.
      operationId: createUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "200":
          description: Created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request payload.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Resource already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        error:
          type: string
        detail:
          type: string
...
    UserRole:
      type: string
      enum: [admin, doctor, specialist, coordinator]
    UserCreate:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        role:
          $ref: "#/components/schemas/UserRole"
        organization_id:
          type: integer
          nullable: true
    UserRead:
      type: object
      required: [id, email, role, is_active]
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
        is_active:
          type: boolean
        organization_id:
          type: integer
          nullable: true
        fhir_id:
          type: string
          nullable: true
...